refactor(memory): split QMD manager responsibilities
